Move transformation fabric functions inside matrix types
